+++ Process.java
@@ -3,40 +3,33 @@
  */
 public class Process {
 
-    private String pid;
-    private int tiempoLlegada;
-    private int tiempoCPU;
-    private int tiempoIO;
+    private int pid;
+    private double tiempo_llegada;
+    private double tiempo_cpu;
     private int prioridad;
-    private int vRuntime;
+    private double v_runtime;
 
 
-    public Process(String pid_, int tiempo_llegada_, int tiempo_io_, int tiempo_cpu_, int prioridad_, int vRuntime){
+    public Process(int pid_, double tiempo_llegada_, double tiempo_cpu_, int prioridad_, double vruntime){
         this.pid = pid_;
-        this.tiempoLlegada = tiempo_llegada_;
-        this.tiempoIO = tiempo_io_;
-        this.tiempoCPU = tiempo_cpu_;
+        this.tiempo_llegada = tiempo_llegada_;
+        this.tiempo_cpu = tiempo_cpu_;
         this.prioridad = prioridad_;
-        this.vRuntime = vRuntime;
+        this.v_runtime = vruntime;
     }
 
-
-    public String get_pid(){
+    public int get_pid(){
         return pid;
     }
 
 
-    public int get_tiempo_llegada(){
-        return tiempoLlegada;
-    }
-
-    public int get_tiempo_io(){
-        return tiempoIO;
+    public double get_tiempo_llegada(){
+        return tiempo_llegada;
     }
 
 
-    public int get_tiempo_cpu(){
-        return tiempoCPU;
+    public double get_tiempo_cpu(){
+        return tiempo_cpu;
     }
 
 
@@ -44,7 +37,7 @@ public class Process {
         return prioridad;
     }
 
-    public int get_vruntime(){
-        return vRuntime;
+    public double get_vruntime(){
+        return v_runtime;
     }
 }
